Consolidate service emulators for tests
Fixes #688

From: Seangchan Ryu <sryu@fnal.gov>


git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/WMCore/trunk@10930 4525493e-7705-40b1-a816-d608a930855b

+++ src/python/WMQuality/Emulators/EmulatorSetup.py
@@ -4,6 +4,7 @@ Use this for only unit test
 import os
 import tempfile
 import sys
+import logging
 
 from WMCore.Configuration import Configuration, saveConfigurationFile
     
@@ -14,7 +15,10 @@ def emulatorSetup(phedex=False, dbs=False, siteDB=False, requestMgr=False):
     return configFile
     
 def deleteConfig(configFile):
-    os.remove(configFile)
+    if os.path.exists(configFile):
+        os.remove(configFile)
+    else:
+        pass
         
 def _emulatorConfig(phedex, dbs, siteDB, requestMgr, configFile):
     
@@ -25,8 +29,9 @@ def _emulatorConfig(phedex, dbs, siteDB, requestMgr, configFile):
     config.Emulator.RequestMgr = requestMgr
     config.Emulator.SiteDB = siteDB
     saveConfigurationFile(config, configFile)
-    print "create config file:%s, PhEDEx: %s, DBS: %s, RequestManager: %s, SiteDB %s with flag" \
+    msg = "create config file:%s, PhEDEx: %s, DBS: %s, RequestManager: %s, SiteDB %s with flag" \
            % (configFile, phedex, dbs, siteDB, requestMgr)
+    logging.info(msg)
            
 def setupWMAgentConfig():
     fd,configFile = tempfile.mkstemp(".py", "TESTAGENTConfig",)
